refacto log-agent
